feat(core): selectable Scalar precision via SE_SCALAR_DOUBLE
Make float vs double a single compile-time build choice, threaded through
the one value-type seam.

- cmake/ProjectOptions.cmake: SE_SCALAR_DOUBLE option (OFF = float).
- CMakeLists.txt: compile definition on the SushiEngine INTERFACE target so
  every consumer (sandbox, pgs_demo, sushi_sim -> editor, tests) agrees on
  sizeof(Scalar); mirrored on sushi_render, which shares the value types
  across its interface structs without linking the engine target.
- core/blas_placeholder.hpp: the sole reader; Float is double under the flag.
- Vulkan upload (vulkan_scene_view.cpp): MeshPush.mvp is an explicit float[16]
  and make_push narrows at the boundary, so GPU data and shaders are unchanged
  in either build.
- Editor: de-hardcode float where it broke the double build — the gizmo
  projection computes in Scalar and casts only the screen coords; scene_camera
  lens params/members are Scalar; inspector widget buffers cast explicitly.
- CLI: --double on `se build` / `se editor`, a persisted scalar_double config
  field, and explicit -DSE_SCALAR_DOUBLE={ON|OFF} that re-caches on toggle.

Both precisions build clean; sandbox validates identically under each
(compile_count==1, mismatches==0) and render_probe stays RESULT: OK.

Docs: CHANGELOG, ARCHITECTURE Â§6, README updated in this commit.

+++ editor/editor_panels.cpp
@@ -386,11 +386,17 @@ namespace sushi::editor
         if (ImGui::CollapsingHeader("Transform", ImGuiTreeNodeFlags_DefaultOpen))
         {
             EntityTransform transform = world->transform(id);
-            float position[3] = {transform.position.x, transform.position.y,
-                                 transform.position.z};
+            // ImGui edits at float precision; the components are engine Scalar (float
+            // or double per build), so narrow explicitly into the widget buffers and
+            // widen back on write.
+            float position[3] = {static_cast<float>(transform.position.x),
+                                 static_cast<float>(transform.position.y),
+                                 static_cast<float>(transform.position.z)};
             float rotation[3];
             quat_to_euler_degrees(transform.rotation, rotation);
-            float scale[3] = {transform.scale.x, transform.scale.y, transform.scale.z};
+            float scale[3] = {static_cast<float>(transform.scale.x),
+                              static_cast<float>(transform.scale.y),
+                              static_cast<float>(transform.scale.z)};
 
             bool changed = false;
             if (ImGui::BeginTable("transform", 2, ImGuiTableFlags_SizingStretchProp))
@@ -415,7 +421,8 @@ namespace sushi::editor
         if (ImGui::CollapsingHeader("Renderer", ImGuiTreeNodeFlags_DefaultOpen))
         {
             const SushiEngine::Vec3 current = world->color(id);
-            float color[3] = {current.x, current.y, current.z};
+            float color[3] = {static_cast<float>(current.x), static_cast<float>(current.y),
+                              static_cast<float>(current.z)};
             if (ImGui::ColorEdit3("Color", color))
                 world->set_color(id, SushiEngine::Vec3{color[0], color[1], color[2]});
         }
